split hooks.ts into multiple files
this avoids a problem where node.js bin scripts would inadevertently import
react-native and crash. splitting the file means that only the hook really
needed in node.js is imported.
