+++ projects/app/src/app/(sidenav)/explore/(words)/words/index.tsx
@@ -1,4 +1,4 @@
-import { useLocalQuery } from "@/client/hooks";
+import { useLocalQuery } from "@/client/hooks/useLocalQuery";
 import { RectButton2 } from "@/client/ui/RectButton2";
 import {
   allHsk1HanziWords,
