+++ projects/app/src/client/hooks/useHanziWordMeaning.ts
@@ -1,0 +1,11 @@
+import type { HanziWord } from "@/data/model";
+import { lookupHanziWord } from "@/dictionary/dictionary";
+import { useLocalQuery } from "./useLocalQuery";
+
+export const useHanziWordMeaning = (hanziWord: HanziWord) => {
+  return useLocalQuery({
+    queryKey: [useHanziWordMeaning.name, hanziWord],
+    queryFn: () => lookupHanziWord(hanziWord),
+    staleTime: Infinity,
+  });
+};
