+++ projects/app/src/client/ui/CircleButton.tsx
@@ -4,7 +4,7 @@ import type { ElementRef } from "react";
 import { forwardRef, useMemo } from "react";
 import type { ColorValue } from "react-native";
 import { Pressable, View } from "react-native";
-import { hapticImpactIfMobile } from "../hooks";
+import { hapticImpactIfMobile } from "../hooks/hapticImpactIfMobile";
 import type { PropsOf } from "./types";
 
 export type CircleButtonProps = {
