+++ projects/app/src/client/ui/PageSheetModal.tsx
@@ -13,7 +13,7 @@ import Reanimated, {
   withSpring,
   withTiming,
 } from "react-native-reanimated";
-import { useEventCallback } from "../hooks";
+import { useEventCallback } from "../hooks/useEventCallback";
 import { AnimatedPressable } from "./AnimatedPressable";
 
 interface PageSheetModalProps {
