+++ projects/app/src/client/ui/QuizProgressBar.tsx
@@ -12,7 +12,7 @@ import Reanimated, {
   withDelay,
   withTiming,
 } from "react-native-reanimated";
-import { useEventCallback } from "../hooks";
+import { useEventCallback } from "../hooks/useEventCallback";
 import type { Transform1D } from "./animate";
 import { createAffineTransform } from "./animate";
